Split the legacy action creation and binding
Previously there was a single big function in BaseInput that would both create
and suggest binds for the legacy input actions. This patch separates those out,
leaving the action creation in BaseInput and moving the bindings out to the
input profiles, which ultimately know best as to where the actions should be
bound on their respective controllers.