Fix invalid layer next pointers in overlays
Previously we set each overlay's next pointer to the next overlay. This
isn't how you're supposed to supply multiple overlays (instead you pass
an array of pointers to the overlays, which is the layerHeaders vector).

This code was lying around from when the code was first written, and
casues validation errors. Hence it's being removed.