BaseInput: Make a registry for actions, actionsets
This implements a new 'registry' system for actions and actionsets. When
an application retrieves an action or actionset, the value they get is a
bit like an XrPath: it's supposed to always come back, even for invalid
actions, and always come back as the same value when looked up.

This patch implements a new generic class for tracking these handles,
and making it easy to convert a handle into the actual data we care
about, if that exists.