Add the OPENCOMPOSITE_DEFAULT_MANIFEST envvar
This makes OpenComposite read the OPENCOMPOSITE_DEFAULT_MANIFEST
environment variable if the game has not loaded an action manifest
before submitting it's first frame. If that happens OpenComposite will
normally assume the game is using the legacy input system, but by
setting the environment variable you can force a given action manifest
to be loaded instead.

Unfortunately we can't change inputs later - OpenXR only lets us set
them once, and then that's it. This is enforced by
xrAttachSessionActionSets which can only be called once in a session's
lifetime.

An example of a game where this is necessary is Jet Island.

This isn't exactly a user-friendly feature, but it's there to allow Jet
Island to be tested before a proper solution to this is found.