Load the empty manifest when legacy input is used
This essentially removes the last commit, by removing the
OPENCOMPOSITE_DEFAULT_MANIFEST environment variable.

Instead, we load the empty manifest on the first call to
BaseSystem::GetControllerState (which is the legacy input function)
after the first frame has been submitted.

This is a rather elegant solution to the problem of when to load the
legacy manifest, since it'll only be loaded if an application is
actually using the legacy input functions (games using the new input
system shouldn't) and don't load their manifests prior to the first
frame (most games load them prior to the first frame).

For games that use legacy input, this doesn't pose a problem: unless
they never poll for input after the first frame (I'd be very interested
to see a game like that) they'll load the empty input manifest almost
immediately.