Initialise all the interfaces at once
Previously, interfaces (referring to the base implementations here, not
the version-specific wrappers which aren't affected by this) were
initialised one at a time, if and when the game requested them.

This wasn't really ideal since as the various interfaces need to
interface with each other more (particularly surrounding input in
OpenXR and the various knock-on effects of trying to support multiple
devices, particularly involving controller properties) this meant that
each interface point requred a check if that interface was loaded.

What happened if an interface wasn't loaded was almost certainly not
comprehensively tested, as most (but certainly not all!) games load
their interfaces without doing anything in-between, but having this
nasty little timebomb sitting around is not a particularly good idea.

Therefore, this patch initialises all the interfaces at once on startup.
Most code can just directly use other interfaces (now calling
GetBaseWhatever instead of GetUnsafeBaseWhatever, since calling it is
now safe in that it's never pointing to a null or just-initialised
object), but a few bits in DrvOpenXR still need to check whether the
interfaces have been initialised, if they can run before that point.

This should also make debugging with the test suite easier, since
stepping into a call doesn't require skipping over all the shared
pointer stuff anymore.