Fix several BaseSystem methods referring to missing controllers
Previously, several methods[1] would return the device IDs of the hand
controllers when they were marked as disconnected, when the type of HMD
hadn't yet been determined.

This patch removes the knowledge that BaseSystem has of the hand
controller IDs, and queries functions of the ITrackedDevice instances
instead. This is both arguably cleaner (less places the hand controller
IDs are hardcoded to) and will make any future support for tracking
pucks slightly easier. Most importantly, it ensures this kind of error
cannot happen as the hands would just be nullptr.

[1] Those methods are:
  * GetTrackedDeviceIndexForControllerRole
  * GetControllerRoleForTrackedDeviceIndex
  * GetTrackedDeviceClass