Return guessed HMD properties before the input profile is known
Currently, most HMD properties are unavailable before the input system
has been loaded. This patch fixes that, by guessing the interaction
profile based on the runtime name.

This reads the OpenXR instance properties and inspects the runtime name.
It uses this to find what interaction profile to return properties from,
until the real one is known.