Consolodate the swapchain waiting code, and handle timeouts
Previously we'd abort if xrWaitSwapchainImage timed out after 500ms.
This patch will log if that happens, but keep retrying until it
succeeds.

This was observed as a problem in DCS: World with WMR, likely because
DCS is apparently very CPU-intesnsive in it's loading screen and was
preventing the WMR runtime from performing frame interpolation in time,
which was preventing it from releasing the previous swapchain images.