Auto-enable invertUsingShaders on Oculus systems
Previously this had to be enabled manually. Since Oculus appears to be
the main runtime which is nonconformant with the inverted FOV settings
(where swapping XrFovf.angleUp and XrFovf.angleDown doesn't invert the
image), this enables the shader-based inverting automatically.

It can still be overridden to either on or off using the settings.