Implement audio override
Added new configuration values for enableAudioSwitch and audioDeviceName.  enableAudioSwitch is a boolean that will cause OpenComposite to read from the string audioDeviceName configuration item and attempt to switch the computer's default audio output device to it.  This is needed so that Oculus headsets don't have to manually set their output when entering skyrimvr.